rename read/*reader.py files for consistency
