+++ anesthetic/read/polychordreader.py
@@ -1,6 +1,6 @@
 """Tools for reading from polychord chains files."""
 import numpy
-from anesthetic.read.getdist import GetDistReader
+from anesthetic.read.getdistreader import GetDistReader
 
 
 class PolyChordReader(GetDistReader):
